bug: 수정
-@Transactional 제거
-entity 기존대로 수정

FDORA-47

+++ src/main/java/com/farmdora/farmdoraauth/service/UserSignUpService.java
@@ -77,7 +77,7 @@ public class UserSignUpService {
                 .email(userSignUpDto.getEmail())
                 .accountNum(userSignUpDto.getAccountNum())
                 .birth(userSignUpDto.getBirth())
-                .sex(userSignUpDto.getSex().getValue())
+                .sex((byte) userSignUpDto.getSex().getValue())
                 .phoneNum(userSignUpDto.getPhoneNum())
                 .bankType(bankType)
                 .auth(Auth.builder()
